Expanded 227 Teal Mushroom to 227 Dye Plant (8 Plants)
Added UV coords for all the Dye Plants as well as changed the size to 1,1 as that is the correct size and prevents the placed items from breaking loose on world load.

The appearance and placement is a bit off in TEdit, but as long as you place based on the mouse pointer position and not the sprite position the placement works fine.

All the plants save Pink Prickly Pear and Orange Bloodroot can be planted on a dirt floor. Orange Bloodroot requires a dirt ceiling while Pink Prickly Pear requires being placed on top of a cactus.

Changed the Color to #FFFFFFFF to aid in easy spotting for future work on this issue.