Colors, formatting, typos, etc...
Mostly formatting fix up, major changes were replacing tabs with spaces, as well as aligning the first entries after ' <Frame UV="xxx,yyy" ' on column 31 (much as ' Color="#FF000000" ' is aligned on column 55). It makes for a much cleaner read.

Non-boring changes are colors added for new tiles/walls as well as tweaked colors for the dungeon walls for better legibility along with Obsidian Brick as it was rather dark against the Underworld background.